+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - [\#39] Add a way of specifying extra GET parameters with the model classes' getLinkedEntity() method
+- [\#39] Add paging parameters to Folder::getCaseFiles()
 
 ### Changed